+++ .env.example
@@ -21,13 +21,20 @@ ACTIX_KEEP_ALIVE=120
 MAX_DB_CONNECTION_POOL_SIZE=3
 TENANT_MIDDLEWARE_EXCLUSION_LIST="/health,/assets/favicon.ico,/pkg/frontend.js,/pkg,/pkg/frontend_bg.wasm,/pkg/tailwind.css,/pkg/style.css,/assets,/admin,/oidc/login,/admin/organisations,/organisations,/organisations/switch/{organisation_id},/"
 SERVICE_PREFIX=""
+### Env(s) for NO AuthN
 AUTH_PROVIDER=DISABLED
+### Env(s) for Simple OIDC AuthN
 # AUTH_PROVIDER=OIDC+http://localhost:8081/realms/users
 # OIDC_CLIENT_ID=superposition
 # OIDC_CLIENT_SECRET=superposition_secret
-# OIDC_TOKEN_ENDPOINT_FORMAT="http://localhost:8081/realms/<organisation>/protocol/openid-connect/token"
-# OIDC_ISSUER_ENDPOINT_FORMAT="http://localhost:8081/realms/<organisation>"
 # OIDC_REDIRECT_HOST="http://localhost:8080"
+### Env(s) for SaaS OIDC AuthN
+# AUTH_PROVIDER=OIDC_SAAS+http://localhost:8081/realms/users
+# OIDC_CLIENT_ID=superposition
+# OIDC_CLIENT_SECRET=superposition_secret
+# OIDC_REDIRECT_HOST="http://localhost:8080"
+# OIDC_ORG_TOKEN_ENDPOINT_FORMAT="http://localhost:8081/realms/<organisation>/protocol/openid-connect/token"
+# OIDC_ORG_ISSUER_ENDPOINT_FORMAT="http://localhost:8081/realms/<organisation>"
 WORKER_ID=1
 ENCRYPTED_KEYS="" # Used for webhook secrets, API keys, etc.
 # ENCRYPTED_KEYS="TEST_API_KEY,DASHBOARD_API_KEY"
