+++ crates/service_utils/src/middlewares/auth_n/no_auth.rs
@@ -1,18 +1,14 @@
-use actix_web::{
-    error,
-    web::{Data, Json},
-    HttpRequest, HttpResponse, Scope,
-};
-use diesel::{Connection, ExpressionMethods, QueryDsl, RunQueryDsl};
+use actix_web::{error, HttpRequest, HttpResponse, Scope};
 use futures_util::future::LocalBoxFuture;
-use superposition_types::{
-    database::superposition_schema::superposition::organisations, User,
-};
+use superposition_types::User;
 
-use crate::service::types::AppState;
+use crate::middlewares::auth_n::helpers::fetch_org_ids_from_db;
 
 use super::authentication::{Authenticator, Login};
 
+/// An Authenticator implementation that performs no authentication
+/// This is primarily for development and testing purposes
+/// In production, a proper Authenticator (like OIDCAuthenticator) should be used
 pub struct DisabledAuthenticator {
     path_prefix: String,
 }
@@ -41,45 +37,8 @@ impl Authenticator for DisabledAuthenticator {
     }
 
     fn get_organisations(&self, req: &actix_web::HttpRequest) -> HttpResponse {
-        let app_state = match req.app_data::<Data<AppState>>() {
-            Some(state) => state,
-            None => {
-                log::info!(
-                    "DbConnection-FromRequest: Unable to get app_data from request"
-                );
-                return error::ErrorInternalServerError(
-                    "Unable to get app_data from request",
-                )
-                .into();
-            }
-        };
-
-        let result = match app_state.db_pool.get() {
-            Ok(mut conn) => {
-                conn.set_prepared_statement_cache_size(
-                    diesel::connection::CacheSize::Disabled,
-                );
-                let orgs = organisations::table
-                    .order(organisations::created_at.desc())
-                    .select(organisations::id)
-                    .get_results::<String>(&mut conn);
-
-                match orgs {
-                    Ok(orgs) => Ok(orgs),
-                    Err(e) => {
-                        log::error!("Failed to fetch organisations: {:?}", e);
-                        Err("Failed to fetch organisations")
-                    }
-                }
-            }
-            Err(e) => {
-                log::info!("Unable to get db connection from pool, error: {e}");
-                Err("Unable to get db connection from pool")
-            }
-        };
-
-        match result {
-            Ok(resp) => HttpResponse::Ok().json(Json(resp)),
+        match fetch_org_ids_from_db(req) {
+            Ok(resp) => HttpResponse::Ok().json(resp),
             Err(resp) => error::ErrorInternalServerError(resp).into(),
         }
     }
